requirements.txt
streamlit
pandas
numpy
opencv-python
pillow
plotly
torch
torchvision
scikit-learn